% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{buildRIVPACS}
\alias{buildRIVPACS}
\alias{getSamples.rivpacs}
\alias{getSites.rivpacs}
\alias{getCalibrationData}
\alias{getGroupMeans}
\alias{getGroupInvCov}
\alias{getCalibrationGroups}
\alias{getCalibrationVars}
\alias{getFormula}
\alias{getModel}
\alias{getCalibrationData.rivpacs}
\alias{getGroupMeans.rivpacs}
\alias{getGroupInvCov.rivpacs}
\alias{getCalibrationGroups.rivpacs}
\alias{getCalibrationVars.rivpacs}
\alias{getFormula.rivpacs}
\alias{getModel.rivpacs}
\title{Build a RIVPACS model}
\usage{
buildRIVPACS(formula, data, type = c("lda", "multinom", "rf"), ...)

\method{getSamples}{rivpacs}(x, ...)

\method{getSites}{rivpacs}(x, ...)

getCalibrationData(x, ...)

getGroupMeans(x, ...)

getGroupInvCov(x, ...)

getCalibrationGroups(x, ...)

getCalibrationVars(x, ...)

getFormula(x, ...)

getModel(x, ...)

\method{getCalibrationData}{rivpacs}(mod)

\method{getGroupMeans}{rivpacs}(mod)

\method{getGroupInvCov}{rivpacs}(mod)

\method{getCalibrationGroups}{rivpacs}(mod)

\method{getCalibrationVars}{rivpacs}(mod)

\method{getFormula}{rivpacs}(mod)

\method{getModel}{rivpacs}(mod)
}
\arguments{
\item{formula}{a formula with the groups on the left and the predictors on the right}

\item{data}{the data frame of site/sample data}

\item{type}{the type of model to build}

\item{\ldots}{additional arguments to be passed to the modeling functions}
}
\value{
a RIVPACS model object (a function)
}
\description{
This function builds a RIVPACS model using either linear discriminant analysis,
random forests, or a multinomial log linear model.  The get* functions can be
used to extract model components.
}
