% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodiversity.R
\name{collapseTaxa}
\alias{collapseTaxa}
\title{Collapse sample data up to a list of higher level taxa.}
\usage{
collapseTaxa(x, taxa)
}
\arguments{
\item{sample}{a vector of sample/site indicatros}

\item{taxon}{a vector of taxa indicators}
}
\description{
Collapses the sample data in a biodiversity object up to a list of higher level taxa. For example,
this allows sample data of multiple taxonomic resolutions to be integrated.
}
