% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIVPACS.R
\name{rarify}
\alias{rarify}
\title{Subsample samples of organisms to a fixed count}
\usage{
rarify(sample, taxon, count, n)
}
\arguments{
\item{sample}{a vector of sample identifiers}

\item{taxon}{a vector of taxon identifiers}

\item{count}{a vector of counts for each taxa in each sample}

\item{n}{the size of the subsample to select}
}
\description{
Subsample samples of organizms to a fixed count.
}
\examples{
d <- expand.grid(sample = 1:2, 
                taxon = c('Ephemeroptera', 'Plecoptera', 'Trichoptera'))
d$count <- runif(6, 10, 20)
rarify(d$sample, d$taxon, d$count, 30)
}
