% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIVPACS.R
\name{resample}
\alias{resample}
\title{Resample a vector}
\usage{
resample(x, size = length(x), ...)
}
\arguments{
\item{x}{a vector}

\item{n}{sample size}
}
\value{
a vector of length n with the sample
}
\description{
This function adapted from the resample function in \code{?sample}. It silently
returns the whole vector in random order when size > length(x).
}
